requirements.txt
cianparser==1.0.4
pandas==2.2.3
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
passlib==1.7.4
starlette-admin==0.14.1
itsdangerous==2.2.0